backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
openai
sendgrid
python-multipart
python-dateutil
pytz
python-jose[cryptography]
passlib[bcrypt]
python-multipart